Up to 24017
The pseudocomment (*=c-*) is needed around headers of large procedures to enable scoping of labels in the assembly code to side-step the 2000 global labels limit; that pseudo-comment at the beginning of the program automatically adds  values from 1 to 47 to the constant pool. 
There seems to be an issue with the source size exceeding 20 (024) zones (120 kBytes). Thus to compile the code successfully, the spaces must be compressed.