Create lazy.pas (программа "Лентяй")
An automated revision tracking tool. If the second line of a source file had the form
(* DD.MM.YY HH.MM MAXADR=NNNNN *)
with digits in place of D, M, Y, H, M, N, and the pseudo-comment Л+ was in effect at the end of the program,
the source file was modified with compilation data/time and the size of the code section if the compilation was successful.
The same date/time were put in the header of the binary.